source java files for the project

+++ Project for traineesheep.String calculator/StringCalculator/src/main/java/ua/goit/Calculator.java
@@ -3,6 +3,8 @@ package ua.goit;
 
 import ua.goit.actions.Action;
 
+import java.math.BigDecimal;
+
 /**
  * This interface can use any way of creation string of formula to make calculations.
  *
@@ -11,7 +13,7 @@ public interface Calculator {
 
     public String getAnswer(String formula) throws Exception;
 
-    public double calculateAnswer(String formula) throws Exception;
+    public BigDecimal calculateAnswer(String formula) throws Exception;
 
     public void setAction(String action, Action newAction);
