+++ Project for traineesheep.String calculator/StringCalculator/src/main/java/ua/goit/actions/Action.java
@@ -1,10 +1,12 @@
 package ua.goit.actions;
 
+import java.math.BigDecimal;
+
 public interface Action {
 
-    public double unaryAction(double argument1);
+    public BigDecimal unaryAction(BigDecimal argument1);
 
-    public double binaryAction(double argument1, double argument2);
+    public BigDecimal binaryAction(BigDecimal argument1, BigDecimal argument2);
 
     public boolean isUnary();
 }
