+++ Project for traineesheep.String calculator/StringCalculator/src/main/java/ua/goit/actions/TanAction.java
@@ -1,5 +1,7 @@
 package ua.goit.actions;
 
+import java.math.BigDecimal;
+
 /**
  * Created by COOLib on 15.05.2016.
  */
@@ -8,13 +10,13 @@ public class TanAction implements Action {
     private boolean isUnary = true;
 
     @Override
-    public double unaryAction(double argument1) {
+    public BigDecimal unaryAction(BigDecimal argument1) {
 
-        return Math.tan(argument1) ;
+        return BigDecimal.valueOf(Math.tan(argument1.doubleValue()));
     }
 
     @Override
-    public double binaryAction(double argument1, double argument2) {
+    public BigDecimal binaryAction(BigDecimal argument1, BigDecimal argument2) {
 
         throw new ArithmeticException("If  the action is tangent, the second argument have to to be here!");
     }
